Working diffuse in compute shader. HORAH

+++ VoxSmith/src/VoxSmith/RayTracer/RayTracer.cpp
@@ -15,7 +15,7 @@ const std::vector<float> g_quadData = {
 	 1,  1, 0,   1, 1, // right-top
 };
 
-RayTracer::RayTracer(const float width, const float height)
+RayTracer::RayTracer(const uint32_t width, const uint32_t height)
 	: m_image(width, height)
 {
 	initBuffer(m_buffer, g_quadData);
